unify names of loggers
The https://docs.python.org/3/library/__main__.html#name-main page says: When a Python module or package is imported, __name__ is set to the module’s name. Usually, this is the name of the Python file itself without the .py extension. If the file is part of a package, __name__ will also include the parent package’s path.

This commit's attempt is to unify logger names so that they all have the same structure, starting with juniorguru. and continuing with inner package and module names, further down to child loggers. This change should prevent having e.g. both "stories" and "juniorguru.sync.stories.blahblah" loggers appearing in a single file.